refactor(form-styles): (reviewed) import input color styles from utility file
